Split XP module from RuneLite module
