+++ src/modules/index.js
@@ -2,11 +2,13 @@ import appReducer from './app'
 import gitReducer from './git'
 import runeliteReducer from './runelite'
 import tagReducer from './tag'
+import xpReducer from './xp'
 
 // Combine all redux reducers into one root reducer
 export default {
   app: appReducer,
   git: gitReducer,
   runelite: runeliteReducer,
-  tag: tagReducer
+  tag: tagReducer,
+  xp: xpReducer
 }
